+++ backend/ddang/src/main/java/com/ddang/ddang/chat/application/ChatRoomService.java
@@ -13,8 +13,8 @@ import com.ddang.ddang.chat.application.exception.ChatRoomNotFoundException;
 import com.ddang.ddang.chat.application.exception.InvalidAuctionToChatException;
 import com.ddang.ddang.chat.application.exception.InvalidUserToChat;
 import com.ddang.ddang.chat.domain.ChatRoom;
-import com.ddang.ddang.chat.infrastructure.persistence.ChatRoomAndMessageRepository;
 import com.ddang.ddang.chat.infrastructure.persistence.JpaChatRoomRepository;
+import com.ddang.ddang.chat.infrastructure.persistence.QuerydslChatRoomAndMessageRepository;
 import com.ddang.ddang.chat.infrastructure.persistence.dto.ChatRoomAndMessageDto;
 import com.ddang.ddang.user.application.exception.UserNotFoundException;
 import com.ddang.ddang.user.domain.User;
@@ -34,7 +34,7 @@ public class ChatRoomService {
     private static final Long DEFAULT_CHAT_ROOM_ID = null;
 
     private final JpaChatRoomRepository chatRoomRepository;
-    private final ChatRoomAndMessageRepository chatRoomAndMessageRepository;
+    private final QuerydslChatRoomAndMessageRepository querydslChatRoomAndMessageRepository;
     private final JpaUserRepository userRepository;
     private final JpaAuctionRepository auctionRepository;
 
@@ -84,7 +84,7 @@ public class ChatRoomService {
         final User findUser = userRepository.findById(userId)
                                             .orElseThrow(() -> new UserNotFoundException("사용자 정보를 찾을 수 없습니다."));
         final List<ChatRoomAndMessageDto> chatRoomAndMessageQueryProjectionDtos =
-                chatRoomAndMessageRepository.findAllChatRoomInfoByUserIdOrderByLastMessage(findUser.getId());
+                querydslChatRoomAndMessageRepository.findAllChatRoomInfoByUserIdOrderByLastMessage(findUser.getId());
 
         return chatRoomAndMessageQueryProjectionDtos.stream()
                                                     .map(dto -> ReadChatRoomWithLastMessageDto.of(findUser, dto))
